# NAVAND-2991 - Optimized TextToSpeech usage