--- a/changelog/unreleased/bugfixes/NAVAND-2991.md
+++ b/changelog/unreleased/bugfixes/NAVAND-2991.md
@@ -1,0 +1,1 @@
+- Fixed UI jank caused by on-device TextToSpeech player.