# Issue

NAVAND-2991 

# Description

Backporting v3 optimization https://github.com/mapbox/navigation/pull/5638

NavSDK 2.0 traces captured from  "examples > Voice Example" activity running on Android Emulator.

| Call | Baseline | Optimized |
|--------|--------|--------|
| `TextToSpeechProvider.getTextToSpeech(...)` | MAIN<br/>`21ms` | **DEFAULT**<br/>`30ms` |
| `VoiceInstructionsTextPlayer.initializeWithLanguage(...)` | MAIN<br/>`207ms` | **DEFAULT**<br/>`508ms` | 
| `textToSpeech.speak(...)` | MAIN<br/>`25ms`<br/>`1ms` | **DEFAULT**<br/>`15ms`<br/>`1ms`| 

I've moved both `initializeWithLanguage(...)` and `textToSpeech.speak(...)` calls to the background thread to optimize this. Additionally, I've updated the "examples > Voice Example" activity to initialize `MapboxVoiceInstructionsPlayer ` off the main thread. The optimized execution time slightly increased due to lesser background thread prioritization.

[navsdk-cpu-perfetto.baseline.trace.zip](https://github.com/user-attachments/files/15920194/navsdk-cpu-perfetto.baseline.trace.zip)
[navsdk-cpu-perfetto.optimized.trace.zip](https://github.com/user-attachments/files/15920195/navsdk-cpu-perfetto.optimized.trace.zip)